#8 - 배포를 위한 레거시 코드 삭제
- 기존에 모든 로직을 새로 제작하고자 기존 파일을 모두 삭제
- 처음부터 다시 시작하기로 함
