--- PokeDex_Project/Model/PokemonDetail.swift
@@ -1,15 +1,0 @@
-//
-//  PokemonDetail.swift
-//  PokeDex_Project
-//
-//  Created by 유영웅 on 2023/08/23.
-//
-
-import Foundation
-
-struct PokemonDetail{
-    var id:Int
-    var name:String
-    var sprite:String
-    var evolDetail:[String]?
-}
