+++ PokeDex_Project/PokeDex_ProjectApp.swift
@@ -6,26 +6,13 @@
 //
 
 import SwiftUI
-//import FirebaseCore
-
-class AppDelegate: NSObject, UIApplicationDelegate {
-  func application(_ application: UIApplication,
-                   didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey : Any]? = nil) -> Bool {
-//    FirebaseApp.configure()
-    return true
-  }
-}
 
 @main
 struct PokeDex_ProjectApp: App {
-    
-    @UIApplicationDelegateAdaptor(AppDelegate.self) var delegate
-    
     var body: some Scene {
         WindowGroup {
             NavigationStack{
-//                ContentView()
-                TestView()
+                ContentView()
             }
         }
     }
