--- PokeDex_Project/Set/ColorSet.swift
@@ -1,16 +1,0 @@
-//
-//  ColorSet.swift
-//  PokeDex_Project
-//
-//  Created by 유영웅 on 2023/04/27.
-//
-
-import Foundation
-import SwiftUI
-
-extension Color{
-    static func typeColor(types:String)->Color{
-        return Color(types)
-    }
-    static let antiPrimary = Color("antiPrimary")
-}
