--- PokeDex_Project/Set/KeyBoardSet.swift
@@ -1,15 +1,0 @@
-//
-//  KeyBoardSet.swift
-//  PokeDex_Project
-//
-//  Created by 유영웅 on 2023/06/16.
-//
-
-import Foundation
-import SwiftUI
-
-extension UIApplication{
-    func endEditing(){
-        sendAction(#selector(UIResponder.resignFirstResponder), to: nil, from: nil, for: nil)
-    }   //키보드 dismissing
-}
